+++ root/urls.py
@@ -16,10 +16,8 @@ Including another URLconf
 from django.contrib import admin
 from django.http import JsonResponse
 from django.urls import path, include
-from django.views.decorators.csrf import csrf_exempt
-from graphene_django.views import GraphQLView
-from accounts.graph_ql import DRFAuthenticatedGraphQLView
 
+from accounts.graph_ql import DRFAuthenticatedGraphQLView
 
 urlpatterns = [
     ############################
@@ -47,6 +45,5 @@ urlpatterns = [
     ############################
     # GRAPH-QL URLS
     ############################
-
     path("graph-ql", DRFAuthenticatedGraphQLView.as_view(graphiql=True))
 ]
